+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Breadcrumbs [#25](https://github.com/azavea/fb-gender-survey-dashboard/pull/25)
+- Add Stack Chart [#26](https://github.com/azavea/fb-gender-survey-dashboard/pull/26)
 
 ### Changed